web/requirements.txt
Django==2.1
djangorestframework==3.9.0
psycopg2==2.7.6.1
pika==0.12.0
django-chartjs==1.3
arrow==0.12.1